refactor of package.json
