+++ app/contact/page.tsx
@@ -1,60 +1,16 @@
-"use client";
-
 import SectionTitle from "@/components/SectionTitle";
 import DarkContainer from "@/components/containers/Dark";
 import LightContainer from "@/components/containers/Light";
+import ContactForm from "@/components/ContactForm";
 import Link from "next/link";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import * as z from "zod";
 
 // SVGR
 import LinkedIn from "@/public/social-icons/linkedin.svg";
 import Github from "@/public/social-icons/github.svg";
 import Phone from "@/public/web-icons/phone.svg";
 import Email from "@/public/web-icons/email.svg";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { Button } from "@/components/ui/button";
-import { Textarea } from "@/components/ui/textarea";
-
-const formSchema = z.object({
-  name: z.string().min(2, {
-    message: "Name must be at least 2 characters",
-  }),
-  email: z.string().email({
-    message: "Must be a valid email address",
-  }),
-  project: z
-    .string()
-    .min(2)
-    .max(500, { message: "No more than 500 characters " }),
-  contact: z.string().min(2, {
-    message: "Name must be at least 2 characters",
-  }),
-});
 
 const ContactPage = () => {
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      name: "",
-      email: "",
-      project: "",
-      contact: "",
-    },
-  });
-
-  function onSubmit(values: z.infer<typeof formSchema>) {
-    console.log(values);
-  }
   return (
     <>
       <DarkContainer>
@@ -115,93 +71,7 @@ const ContactPage = () => {
             </div>
           </aside>
           <article className="w-full md:w-2/3 flex flex-col gap-9 md:gap-20">
-            <Form {...form}>
-              <form
-                onSubmit={form.handleSubmit(onSubmit)}
-                className="space-y-8"
-              >
-                <FormField
-                  control={form.control}
-                  name="name"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel className="text-body-regular md:text-paragraph-regular text-primary-black-300 dark:text-primary-white-800">
-                        What&apos;s your full name?
-                      </FormLabel>
-                      <FormControl className="bg-primary-white-800 h-16 md:h-20">
-                        <Input
-                          {...field}
-                          className="text-body-regular md:text-paragraph-regular rounded-lg"
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="email"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel className="text-body-regular md:text-paragraph-regular text-primary-black-300 dark:text-primary-white-800">
-                        What&apos;s your email?
-                      </FormLabel>
-                      <FormControl className="bg-primary-white-800 h-16 md:h-20">
-                        <Input
-                          {...field}
-                          className="text-body-regular md:text-paragraph-regular rounded-lg"
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="project"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel className="text-body-regular md:text-paragraph-regular text-primary-black-300 dark:text-primary-white-800">
-                        Write something about your project goals and timeframe
-                      </FormLabel>
-                      <FormControl className="bg-primary-white-800 h-24 md:h-40">
-                        <Textarea
-                          {...field}
-                          className="text-body-regular md:text-paragraph-regular rounded-lg"
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="contact"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel className="text-body-regular md:text-paragraph-regular text-primary-black-300 dark:text-primary-white-800">
-                        How can I contact you? ex. phone number or email address
-                      </FormLabel>
-                      <FormControl className="bg-primary-white-800 h-16 md:h-20">
-                        <Input
-                          {...field}
-                          className="text-body-regular md:text-paragraph-regular rounded-lg"
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <div className="flex items-center justify-end">
-                  <Button
-                    type="submit"
-                    className="text-body-bold text-white w-full md:w-52"
-                  >
-                    Send
-                  </Button>
-                </div>
-              </form>
-            </Form>
+            <ContactForm />
           </article>
         </section>
       </LightContainer>
