+++ package.json
@@ -16,6 +16,7 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-switch": "^1.0.3",
     "@radix-ui/react-tooltip": "^1.0.6",
+    "@sanity/vision": "^3.16.7",
     "@types/node": "20.5.9",
     "@types/react": "18.2.21",
     "@types/react-dom": "18.2.7",
@@ -34,6 +35,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.46.2",
+    "sanity": "^3.16.7",
+    "styled-components": "^5.3.11",
     "tailwind-merge": "^1.14.0",
     "tailwindcss": "3.3.3",
     "tailwindcss-animate": "^1.0.7",
@@ -41,6 +44,7 @@
     "zod": "^3.22.2"
   },
   "devDependencies": {
+    "@sanity/eslint-config-studio": "^3.0.1",
     "@svgr/webpack": "^8.1.0"
   }
 }
